📝 Chore: map 관련 타입 지정
