+++ src/types/map.ts
@@ -1,0 +1,35 @@
+interface Coordinates {
+  latitude: number;
+  longitude: number;
+}
+
+interface Position {
+  coords: Coordinates;
+}
+
+interface CustomCoordinates {
+  lat: number;
+  lng: number;
+}
+
+interface UserPosition {
+  userPosition: CustomCoordinates | null;
+}
+
+interface MapMarker {
+  position: {
+    lat: number;
+    lng: number;
+  };
+  content: string;
+  address: string;
+  phone: string;
+}
+
+export type {
+  Coordinates,
+  Position,
+  CustomCoordinates,
+  UserPosition,
+  MapMarker,
+};
